requirements.txt
# Core App & Data
streamlit>=1.35.0
pandas
python-dotenv
pdfplumber

# Browser Automation
selenium
webdriver-manager

# AI & Agents
crewai
langchain-google-genai
google-generativeai
